requirements.txt
aiogram>=3.0.0
python-dotenv>=0.19.0
structlog>=23.1.0
aiohttp>=3.8.0
httpx>=0.24.0
openai>=1.0.0
boto3>=1.28.0
botocore>=1.31.0
tenacity>=8.2.0
prometheus-client>=0.16.0
yadisk>=1.2.0
